// spot-commons/exceptions/messages.d.ts
export declare const ERROR_MESSAGES: {
    RATE_LIMIT: {
        RATE_LIMIT: string;
    };
    MAIN: {
        FRIENDS: {
            GET_FRIENDS: string;
            DELETE_FRIEND: string;
            GET_FRIEND_REQUESTS: string;
            GET_PENDING_FRIEND_REQUESTS: string;
            DELETE_PENDING_FRIEND_REQUEST: string;
            ACCEPT_FRIEND_REQUEST: string;
            DECLINE_FRIEND_REQUEST: string;
            EXISTS: string;
            REQUEST_EXISTS: string;
            SELF: string;
            NO_USER: string;
            GENERIC: string;
        };
        POSTS: {
            POST_ERROR: string;
            INVALID_POST_CONTENT: string;
            INVALID_POST_LENGTH: string;
            INVALID_POST_LINE_LENGTH: string;
            NO_CONTENT: string;
            IMAGE: string;
            INVALID_POST_PROFANITY: string;
            POST_ACTIVITY: string;
            GET_SINGLE_POST: string;
            DELETE_POST: string;
            DISLIKE_POST: string;
            LIKE_POST: string;
            UNRATED_POST: string;
            GET_POSTS: string;
        };
        COMMENTS: {
            ADD_COMMENT: string;
            INVALID_COMMENT_CONTENT: string;
            INVALID_COMMENT_LENGTH: string;
            IMAGE: string;
            INVALID_COMMENT_LINE_LENGTH: string;
            NO_CONTENT: string;
            INVALID_COMMENT_PROFANITY: string;
            COMMENT_ACTIVITY: string;
            GET_COMMENTS: string;
            GET_REPLIES: string;
            DELETE_COMMENT: string;
            DELETE_REPLY: string;
            LIKE_COMMENT: string;
            DISLIKE_COMMENT: string;
            UNRATED_COMMENT: string;
            LIKE_REPLY: string;
            DISLIKE_REPLY: string;
            UNRATED_REPLY: string;
            REPORT_COMMENT: string;
            NOT_IN_RANGE: string;
            NOT_TAGGED: string;
        };
        REPORT: {
            REPORT_ERROR: string;
            REPORT_LENGTH: string;
        };
        LOCATION: {
            LOCATION_ERROR: string;
        };
        NOTIFICATIONS: {
            GET_NOTIFICATIONS: string;
            GET_UNREAD_NOTIFICATIONS: string;
            SEND_NOTIFICATION: string;
            SEEN_NOTIFICATION: string;
            SEEN_ALL_NOTIFICATION: string;
            DELETE_NOTIFICATION: string;
            DELETE_ALL_NOTIFICATION: string;
        };
        ACCOUNTS: {
            DELETE_ACCOUNT: string;
            GET_ACCOUNT: string;
            UPDATE_USERNAME: string;
            UPDATE_USERNAME_TIMEOUT: string;
            UPDATE_EMAIL: string;
            UPDATE_EMAIL_TIMEOUT: string;
            UPDATE_PHONE: string;
            UPDATE_PHONE_TIMEOUT: string;
            FACEBOOK_CONNECT: string;
            FACEBOOK_CONNECT_EXISTS: string;
            FACEBOOK_DISCONNECT: string;
            GOOGLE_CONNECT: string;
            GOOGLE_DISCONNECT: string;
            GET_METADATA: string;
            METADATA_DISTANCE_UNIT: string;
            METADATA_SEARCH_TYPE: string;
            METADATA_SEARCH_DISTANCE: string;
            METADATA_MATURE_FILTER: string;
            METADATA_THEME_WEB: string;
            SEND_VERIFY: string;
            CONFIRM_VERIFY: string;
        };
    };
    PRE_AUTH: {
        AUTHENTICATION: {
            NO_AUTHENTICATION: string;
            VERIFY: string;
            REGISTER: string;
            USERNAME_TAKEN: string;
            USERNAME_CHARACTER: string;
            USERNAME_LENGTH: string;
            PASSWORD_LENGTH: string;
            EMAIL_TAKEN: string;
            EMAIL_INVALID: string;
            PHONE_TAKEN: string;
            PHONE_INVALID: string;
            USERNAME_OR_PASSWORD: string;
            PASSWORD_RESET: string;
            PASSWORD_RESET_VALIDATE: string;
            NEW_PASSWORD: string;
            UPDATE_USERNAME: string;
            FACEBOOK_SIGNUP: string;
            GOOGLE_SIGNUP: string;
        };
    };
};